requirements.txt
psycopg2-binary>=2.7.4
requests>=2.18.4
facebook-sdk>=3.0.0
boto3>=1.6.22